requirements.txt
python-decouple
aiogram
requests
urllib3